% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_ConstructorF.R
\name{createLncRNAGene}
\alias{createLncRNAGene}
\title{Create a Long Non-Coding RNA Gene Object}
\usage{
createLncRNAGene(
  id,
  symbol,
  name,
  description,
  structure,
  lncRNA_id,
  RNA_sequence
)
}
\arguments{
\item{id}{Character. Gene identifier.}

\item{symbol}{Character. Gene symbol.}

\item{name}{Character. Gene name.}

\item{description}{Character. Gene description.}

\item{structure}{GRanges. Gene structure data.}

\item{lncRNA_id}{Character. lncRNA identifier.}

\item{RNA_sequence}{Character. RNA sequence.}
}
\value{
An instance of the lncRNAGene class.
}
\description{
Constructor function for creating an lncRNAGene object.
}
\examples{
library(GenomicRanges)
structure <- GRanges(
  seqnames = "chr11",
  ranges = IRanges(start = 65497758, end = 65503100),
  strand = "+"
)
malat1 <- createLncRNAGene(
  id = "ENSG000002",
  symbol = "MALAT1",
  name = "Metastasis Associated Lung Adenocarcinoma Transcript 1",
  description = "lncRNA associated with metastasis",
  structure = structure,
  lncRNA_id = "NONHSAG000002",
  RNA_sequence = "AUGCUACGUGA"
)
print(malat1)
}
