% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_ConstructorF.R
\name{createMiRNAGene}
\alias{createMiRNAGene}
\title{Create a microRNA Gene Object}
\usage{
createMiRNAGene(
  id,
  symbol,
  name,
  description,
  structure,
  miRNA_id,
  seed_sequence
)
}
\arguments{
\item{id}{Character. Gene identifier.}

\item{symbol}{Character. Gene symbol.}

\item{name}{Character. Gene name.}

\item{description}{Character. Gene description.}

\item{structure}{GRanges. Gene structure data.}

\item{miRNA_id}{Character. microRNA identifier.}

\item{seed_sequence}{Character. Seed sequence of the microRNA.}
}
\value{
An instance of the miRNAGene class.
}
\description{
Constructor function for creating a miRNAGene object.
}
\examples{
library(GenomicRanges)
structure <- GRanges(
  seqnames = "chr21",
  ranges = IRanges(start = 25575833, end = 25575902),
  strand = "-"
)
mir21 <- createMiRNAGene(
  id = "ENSG000003",
  symbol = "MIR21",
  name = "microRNA 21",
  description = "A microRNA associated with cancer",
  structure = structure,
  miRNA_id = "MI0000077",
  seed_sequence = "AGCUUA"
)
print(mir21)
}
