% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{Gene-class}
\alias{Gene-class}
\title{Gene Class}
\arguments{
\item{object}{An object of class Gene.}

\item{value}{A value to be assigned to the slot.}
}
\description{
A virtual S4 class to represent the generalities of genes.
}
\section{Slots}{

\describe{
\item{\code{id}}{Character. The gene identifier (e.g., Ensembl or NCBI gene ID).}

\item{\code{symbol}}{Character. The HUGO symbol of the gene.}

\item{\code{name}}{Character. The full name of the gene.}

\item{\code{description}}{Character. The description of the gene.}

\item{\code{structure}}{GRanges. The genomic structure of the gene
(chromosome, start, end, strand).}
}}

\examples{
library(GenomicRanges)
# Define a generic genomic structure
structure <- GRanges(
  seqnames = "chr1",
  ranges = IRanges(start = 1000, end = 5000),
  strand = "+"
)
# Create a generic gene object
gene <- new(
  "Gene",
  id = "GENE000001",
  symbol = "GENE1",
  name = "Generic Gene",
  description = "This is a generic example of a gene.",
  structure = structure
)
print(gene)
}
\keyword{classes}
