% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_AccessorF.R
\name{GeneAccessors}
\alias{GeneAccessors}
\alias{structure}
\alias{structure,Gene-method}
\alias{structure<-}
\alias{structure<-,Gene-method}
\alias{symbol}
\alias{symbol,Gene-method}
\alias{symbol<-}
\alias{symbol<-,Gene-method}
\alias{name}
\alias{name,Gene-method}
\alias{name<-}
\alias{name<-,Gene-method}
\title{Gene Accessors}
\usage{
structure(object)

\S4method{structure}{Gene}(object)

structure(object) <- value

\S4method{structure}{Gene}(object) <- value

symbol(object)

\S4method{symbol}{Gene}(object)

symbol(object) <- value

\S4method{symbol}{Gene}(object) <- value

name(object)

\S4method{name}{Gene}(object)

name(object) <- value

\S4method{name}{Gene}(object) <- value
}
\arguments{
\item{object}{An object of class Gene or its subclasses.}

\item{value}{A value to set for the respective slot.}
}
\value{
A character string representing the gene symbol.

The updated gene object.

A character string representing the gene name.

The updated gene object.
}
\description{
Provides accessor methods for retrieving
or updating slots in the Gene class and its subclasses.

Retrieves the HUGO symbol of a gene.

Updates the HUGO symbol of a gene.

Retrieves the name of a gene.

Updates the name of a gene.
}
\examples{
library(GenomicRanges)
gene <- createProteinCodingGene(
  id = "ENSG000001",
  symbol = "TP53",
  name = "Tumor Protein",
  description = "Tumor suppressor gene",
  structure = GRanges(
    seqnames = "chr17",
    ranges = IRanges(start = 7668402, end = 7687550),
    strand = "+"
  ),
  protein_id = "P04637",
  protein_sequence = "MVLSPADKTNVKAAWG"
)
symbol(gene)
library(GenomicRanges)
gene <- createProteinCodingGene(
  id = "ENSG000001",
  symbol = "TP53",
  name = "Tumor Protein",
  description = "Tumor suppressor gene",
  structure = GRanges(
    seqnames = "chr17",
    ranges = IRanges(start = 7668402, end = 7687550),
    strand = "+"
  ),
  protein_id = "P04637",
  protein_sequence = "MVLSPADKTNVKAAWG"
)
symbol(gene) <- "TP53_UPDATED"
library(GenomicRanges)
gene <- createProteinCodingGene(
  id = "ENSG000001",
  symbol = "TP53",
  name = "Tumor Protein",
  description = "Tumor suppressor gene",
  structure = GRanges(
    seqnames = "chr17",
    ranges = IRanges(start = 7668402, end = 7687550),
    strand = "+"
  ),
  protein_id = "P04637",
  protein_sequence = "MVLSPADKTNVKAAWG"
)
name(gene)
library(GenomicRanges)
gene <- createProteinCodingGene(
  id = "ENSG000001",
  symbol = "TP53",
  name = "Tumor Protein",
  description = "Tumor suppressor gene",
  structure = GRanges(
    seqnames = "chr17",
    ranges = IRanges(start = 7668402, end = 7687550),
    strand = "+"
  ),
  protein_id = "P04637",
  protein_sequence = "MVLSPADKTNVKAAWG"
)
name(gene) <- "TP53 Protein"
}
