% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{lncRNAGene-class}
\alias{lncRNAGene-class}
\title{Long Non-Coding RNA (lncRNA) Gene Class}
\description{
An S4 class representing long non-coding RNA (lncRNA) genes,
inheriting properties from the Gene class.
}
\section{Slots}{

\describe{
\item{\code{lncRNA_id}}{Character. lncRNA identifier.}

\item{\code{RNA_sequence}}{Character. RNA sequence.}
}}

\examples{
library(GenomicRanges)
structure <- GRanges(
  seqnames = "chr11",
  ranges = IRanges(start = 65497758, end = 65503100),
  strand = "+"
)
malat1 <- createLncRNAGene(
  id = "ENSG000002",
  symbol = "MALAT1",
  name = "Metastasis Associated Lung Adenocarcinoma Transcript 1",
  description = "lncRNA associated with metastasis",
  structure = structure,
  lncRNA_id = "NONHSAG000002",
  RNA_sequence = "AUGCUACGUGA"
)
print(malat1)
}
\keyword{classes}
