% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{miRNAGene-class}
\alias{miRNAGene-class}
\title{microRNA Gene Class}
\description{
An S4 class representing microRNA (miRNA) genes,
inheriting properties from the Gene class.
}
\section{Slots}{

\describe{
\item{\code{miRNA_id}}{Character. microRNA identifier.}

\item{\code{seed_sequence}}{Character. Seed sequence of the microRNA.}
}}

\examples{
library(GenomicRanges)
structure <- GRanges(
  seqnames = "chr21",
  ranges = IRanges(start = 25575833, end = 25575902),
  strand = "-"
)
mir21 <- createMiRNAGene(
  id = "ENSG000003",
  symbol = "MIR21",
  name = "microRNA 21",
  description = "A microRNA associated with cancer",
  structure = structure,
  miRNA_id = "MI0000077",
  seed_sequence = "AGCUUA"
)
print(mir21)
}
\keyword{classes}
