% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{ProteinCodingGene-class}
\alias{ProteinCodingGene-class}
\title{Protein Coding Gene Class}
\description{
An S4 class representing protein-coding genes,
inheriting general properties from the Gene class.
}
\section{Slots}{

\describe{
\item{\code{protein_id}}{Character. Protein identifier.}

\item{\code{protein_sequence}}{Character. Protein sequence.}
}}

\examples{
library(GenomicRanges)
structure <- GRanges(
  seqnames = "chr17",
  ranges = IRanges(start = 7668402, end = 7687550),
  strand = "+"
)
tp53 <- createProteinCodingGene(
  id = "ENSG000001",
  symbol = "TP53",
  name = "Tumor Protein",
  description = "Tumor suppressor gene",
  structure = structure,
  protein_id = "P04637",
  protein_sequence = "MVLSPADKTNVKAAWG"
)
print(tp53)
}
\keyword{classes}
