% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_ConstructorF.R
\name{validateGRanges}
\alias{validateGRanges}
\title{Validate GRanges Structure}
\usage{
validateGRanges(structure)
}
\arguments{
\item{structure}{A GRanges object containing gene structure information.}
}
\value{
Throws an error if the structure is not a GRanges object.
}
\description{
Checks whether the provided structure is a valid GRanges object.
}
\keyword{internal}
